refactor(api): consolidate get_session_and_player into shared middleware
Nearly every route blueprint defined its own copy of
get_session_and_player with minor signature drift (some returned a
5-tuple with a separate status code, others took request as an arg).
Move the canonical implementation into src/api/middleware/auth.py and
route all 16 blueprints plus app.py through it.
