+++ src/api/routes/debug.py
@@ -10,33 +10,9 @@ create_app), so it is never reachable in production.
 """
 
 from flask import Blueprint, jsonify, request
+from src.api.middleware.auth import get_session_and_player
 
 debug_bp = Blueprint("debug", __name__)
-
-
-def get_session_and_player(request):
-    """Resolve (session_manager, session, player, error) from a Bearer token.
-
-    Mirrors the helper used by the other route blueprints.
-    """
-    from flask import current_app
-
-    auth_header = request.headers.get("Authorization", "")
-    if not auth_header.startswith("Bearer "):
-        return None, None, None, (jsonify({"error": "Missing authorization"}), 401)
-
-    session_id = auth_header[7:]
-    session_manager = current_app.session_manager
-    session = session_manager.get_session(session_id)
-    if not session:
-        return None, None, None, (
-            jsonify({"error": "Invalid or expired session"}),
-            401,
-        )
-    player = session_manager.get_player(session_id)
-    if not player:
-        return None, None, None, (jsonify({"error": "Player not found"}), 404)
-    return session_manager, session, player, None
 
 
 _adjutant_instance = None
@@ -54,7 +30,7 @@ def _adjutant():
 
 def _run(operation):
     """Resolve the session player and run `operation(adjutant, player, body)`."""
-    _sm, _session, player, error = get_session_and_player(request)
+    _sm, _session, player, error = get_session_and_player()
     if error:
         return error
     body = request.get_json(silent=True) or {}
