+++ src/api/routes/equipment.py
@@ -1,40 +1,9 @@
 """Equipment management routes."""
 
 from flask import Blueprint, request, jsonify
+from src.api.middleware.auth import get_session_and_player
 
 equipment_bp = Blueprint("equipment", __name__)
-
-
-def get_session_and_player(request):
-    """Extract session and player from request."""
-    from flask import current_app
-
-    auth_header = request.headers.get("Authorization", "")
-    if not auth_header.startswith("Bearer "):
-        return (
-            None,
-            None,
-            None,
-            (jsonify({"error": "Missing authorization"}), 401),
-        )
-
-    session_id = auth_header[7:]
-    session_manager = current_app.session_manager
-    session = session_manager.get_session(session_id)
-
-    if not session:
-        return (
-            None,
-            None,
-            None,
-            (jsonify({"error": "Invalid or expired session"}), 401),
-        )
-
-    player = session_manager.get_player(session_id)
-    if not player:
-        return None, None, None, (jsonify({"error": "Player not found"}), 404)
-
-    return session_manager, session, player, None
 
 
 @equipment_bp.route("/equipment", methods=["GET"])
@@ -58,7 +27,7 @@ def get_equipment():
         }
     """
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -113,7 +82,7 @@ def equip_item():
         }
     """
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -176,7 +145,7 @@ def unequip_item():
         }
     """
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
