+++ src/api/routes/reputation.py
@@ -1,61 +1,10 @@
 """Reputation management routes for Phase 3 Stage 2."""
 
 from flask import Blueprint, request, jsonify, current_app
-from typing import Any, Optional, Tuple
+
+from src.api.middleware.auth import get_session_and_player
 
 reputation_bp = Blueprint("reputation", __name__, url_prefix="/api/reputation")
-
-
-def get_session_and_player(
-    req: Any,
-) -> Tuple[Any, Any, Any, Optional[Any], Optional[int]]:
-    """Extract and validate session from request header.
-
-    Args:
-        req: Flask request object
-
-    Returns:
-        Tuple of (session_manager, session, player, error_response, error_code)
-    """
-    auth_header = req.headers.get("Authorization", "")
-    if not auth_header.startswith("Bearer "):
-        return (
-            None,
-            None,
-            None,
-            jsonify(
-                {
-                    "success": False,
-                    "error": "Missing or invalid authentication header",
-                }
-            ),
-            401,
-        )
-
-    session_id = auth_header[7:]
-    session_manager = current_app.session_manager
-    session = session_manager.get_session(session_id)
-
-    if not session:
-        return (
-            None,
-            None,
-            None,
-            jsonify({"success": False, "error": "Invalid or expired session"}),
-            401,
-        )
-
-    player = session_manager.get_player(session_id)
-    if not player:
-        return (
-            None,
-            None,
-            None,
-            jsonify({"success": False, "error": "Player not found for session"}),
-            404,
-        )
-
-    return session_manager, session, player, None, None
 
 
 @reputation_bp.route("/player", methods=["GET"])
@@ -65,11 +14,9 @@ def get_player_reputation():
     Requires: Bearer token in Authorization header
     Returns: All reputation data
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     result = current_app.game_service.get_player_reputation(player)
     return jsonify({"success": True, "data": result["reputation"]}), 200
@@ -85,11 +32,9 @@ def get_npc_relationship(npc_id: str):
     Requires: Bearer token in Authorization header
     Returns: Relationship data including reputation, attitude, and flags
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     if not npc_id or not isinstance(npc_id, str):
         return (
@@ -121,11 +66,9 @@ def update_npc_relationship(npc_id: str):
     Requires: Bearer token in Authorization header
     Returns: Reputation change details
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     if not npc_id or not isinstance(npc_id, str):
         return (
@@ -196,11 +139,9 @@ def set_relationship_flag(npc_id: str, flag_name: str):
     Requires: Bearer token in Authorization header
     Returns: Flag update result
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     if not npc_id or not isinstance(npc_id, str):
         return (
@@ -255,11 +196,9 @@ def check_dialogue_available(npc_id: str, dialogue_node: str):
     Requires: Bearer token in Authorization header
     Returns: Availability status and lock reason if applicable
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     if not npc_id or not isinstance(npc_id, str):
         return (
@@ -291,11 +230,9 @@ def check_quest_available(npc_id: str, quest_type: str):
     Requires: Bearer token in Authorization header
     Returns: Availability status and lock reason if applicable
     """
-    session_manager, session, player, error, status_code = get_session_and_player(
-        request
-    )
+    session_manager, session, player, error = get_session_and_player()
     if error:
-        return error, status_code
+        return error
 
     if not npc_id or not isinstance(npc_id, str):
         return (
