+++ src/api/routes/saves.py
@@ -1,47 +1,16 @@
 """Save and load game routes."""
 
 from flask import Blueprint, request, jsonify
+from src.api.middleware.auth import get_session_and_player
 
 saves_bp = Blueprint("saves", __name__)
-
-
-def get_session_and_player(request):
-    """Extract session and player from request."""
-    from flask import current_app
-
-    auth_header = request.headers.get("Authorization", "")
-    if not auth_header.startswith("Bearer "):
-        return (
-            None,
-            None,
-            None,
-            (jsonify({"error": "Missing authorization"}), 401),
-        )
-
-    session_id = auth_header[7:]
-    session_manager = current_app.session_manager
-    session = session_manager.get_session(session_id)
-
-    if not session:
-        return (
-            None,
-            None,
-            None,
-            (jsonify({"error": "Invalid or expired session"}), 401),
-        )
-
-    player = session_manager.get_player(session_id)
-    if not player:
-        return None, None, None, (jsonify({"error": "Player not found"}), 404)
-
-    return session_manager, session, player, None
 
 
 @saves_bp.route("/saves", methods=["GET"])
 async def list_saves():
     """List all saved games for player from Turso cloud storage."""
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -73,7 +42,7 @@ async def list_saves():
 async def create_save():
     """Create a new manual or auto save in Turso cloud."""
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -144,7 +113,7 @@ async def create_save():
 async def load_save(save_id):
     """Load a saved game from Turso cloud."""
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -206,7 +175,7 @@ async def load_save(save_id):
 async def delete_save(save_id):
     """Delete a saved game from Turso cloud."""
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
 
@@ -269,7 +238,7 @@ def new_game():
         }
     """
     try:
-        session_manager, session, player, error = get_session_and_player(request)
+        session_manager, session, player, error = get_session_and_player()
         if error:
             return error[0], error[1]
